requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.11.2
tf-nightly
# tensorflow==2.18.0
# keras==3.8.0
joblib
nltk
pillow